+++ code/packages/store-azure-web-pubsub/src/server/azure-web-pubsub-host.ts
@@ -24,11 +24,15 @@ import {
   type WeaveStoreAzureWebPubsubOnWebsocketErrorEvent,
   type WeaveStoreAzureWebPubsubOnWebsocketJoinGroupEvent,
   type WeaveStoreAzureWebPubsubOnWebsocketMessageEvent,
-  type WeaveStoreAzureWebPubsubOnWebsocketOnTokenRefreshEvent,
   type WeaveStoreAzureWebPubsubOnWebsocketOpenEvent,
+  type WeaveStoreAzureWebPubsubOnWebsocketReconnectEvent,
+  type WeaveStoreAzureWebPubsubSyncHostOptions,
 } from '@/types';
 import type WeaveAzureWebPubsubSyncHandler from './azure-web-pubsub-sync-handler';
 import { handleChunkedMessage } from '@/utils';
+import type { DeepPartial } from '@inditextech/weave-types';
+import { mergeExceptArrays } from '@inditextech/weave-sdk';
+import { WEAVE_STORE_AZURE_WEB_PUBSUB_SYNC_HOST_DEFAULT_OPTIONS } from '@/constants';
 
 const expirationTimeInMinutes = 60; // 1 hour
 const messageSync = 0;
@@ -59,6 +63,13 @@ export class WeaveStoreAzureWebPubSubSyncHost {
 
   private _chunkedMessages: Map<string, string[]>;
 
+  private readonly _syncHostOptions: WeaveStoreAzureWebPubsubSyncHostOptions;
+
+  private _heartbeatIntervalId: NodeJS.Timeout | null;
+
+  private _resyncAttempt: number = 0;
+  private _resyncIntervalId: NodeJS.Timeout | null = null;
+
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private _updateHandler: (update: any, origin: any) => void;
   private _awarenessUpdateHandler: (
@@ -73,8 +84,13 @@ export class WeaveStoreAzureWebPubSubSyncHost {
     syncHandler: WeaveAzureWebPubsubSyncHandler,
     client: WebPubSubServiceClient,
     topic: string,
-    doc: Y.Doc
+    doc: Y.Doc,
+    syncHostOptions?: DeepPartial<WeaveStoreAzureWebPubsubSyncHostOptions>
   ) {
+    this._syncHostOptions = mergeExceptArrays(
+      WEAVE_STORE_AZURE_WEB_PUBSUB_SYNC_HOST_DEFAULT_OPTIONS,
+      syncHostOptions ?? {}
+    );
     this.server = server;
     this.syncHandler = syncHandler;
     this.doc = doc;
@@ -82,6 +98,8 @@ export class WeaveStoreAzureWebPubSubSyncHost {
     this.topicAwarenessChannel = `${topic}-awareness`;
     this._client = client;
     this._chunkedMessages = new Map();
+
+    this._heartbeatIntervalId = null;
 
     this._conn = null;
 
@@ -164,20 +182,40 @@ export class WeaveStoreAzureWebPubSubSyncHost {
     this.broadcast(this.topic, origin, u8);
   }
 
+  private setupHeartbeat() {
+    this._heartbeatIntervalId = setInterval(() => {
+      this._conn?.send?.(
+        JSON.stringify({
+          type: MessageType.SendToGroup,
+          group: this.topic,
+          noEcho: true,
+          data: { type: 'heartbeat' },
+        })
+      );
+    }, this._syncHostOptions.heartbeat.sendIntervalMs);
+  }
+
   async createWebSocket(): Promise<void> {
     const group = this.topic;
 
     const { url } = await this.negotiate(this.topic);
 
-    this._reconnectAttempts++;
-
     return new Promise((resolve) => {
       const ws = new WebSocket(url, AzureWebPubSubJsonProtocol);
+
+      const connectionAttempt = this._reconnectAttempts;
+
+      this._resyncAttempt = 0;
+      this._resyncIntervalId = null;
 
       ws.addEventListener('open', (event) => {
         this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketOpenEvent>(
           'onWsOpen',
-          { group: `${group}.host`, event }
+          {
+            group: `${group}.host`,
+            event,
+            connectionAttempt,
+          }
         );
         ws.send(
           JSON.stringify({
@@ -187,8 +225,39 @@ export class WeaveStoreAzureWebPubSubSyncHost {
         );
         this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketJoinGroupEvent>(
           'onWsJoinGroup',
-          { group: `${group}.host` }
+          {
+            group: `${group}.host`,
+            connectionAttempt,
+          }
         );
+
+        const handleResync = () => {
+          ws.send(
+            JSON.stringify({
+              type: MessageType.SendToGroup,
+              group,
+              noEcho: true,
+              data: { type: 'resync' },
+            })
+          );
+          this._resyncAttempt++;
+          if (
+            this._resyncAttempt >= this._syncHostOptions.resync.attemptsLimit &&
+            this._resyncIntervalId
+          ) {
+            clearInterval(this._resyncIntervalId);
+          }
+        };
+
+        this._resyncIntervalId = setInterval(() => {
+          handleResync();
+        }, this._syncHostOptions.resync.checkIntervalMs);
+
+        handleResync();
+
+        this.setupHeartbeat();
+
+        this._reconnectAttempts = 0; // reset on successful connection
 
         this._conn = ws;
         resolve();
@@ -197,7 +266,10 @@ export class WeaveStoreAzureWebPubSubSyncHost {
       ws.addEventListener('message', (e) => {
         this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketMessageEvent>(
           'onWsMessage',
-          { group: `${group}.host`, event: e }
+          {
+            group: `${group}.host`,
+            event: e,
+          }
         );
 
         const event: Message = JSON.parse(e.data.toString());
@@ -238,30 +310,59 @@ export class WeaveStoreAzureWebPubSubSyncHost {
       });
 
       ws.addEventListener('close', (e) => {
+        if (this._heartbeatIntervalId) {
+          clearInterval(this._heartbeatIntervalId);
+        }
+
+        if (this._resyncIntervalId) {
+          clearInterval(this._resyncIntervalId);
+        }
+
         this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketCloseEvent>(
           'onWsClose',
           {
             group: `${group}.host`,
             event: e as unknown as CloseEvent,
+            connectionAttempt,
           }
         );
 
         if (this._forceClose) {
           return;
         } else {
-          const timeout = 1000 * Math.pow(1.5, this._reconnectAttempts);
+          this._reconnectAttempts++;
+          const timeoutMs = 1000 * Math.pow(1.5, this._reconnectAttempts - 1);
+
+          this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketReconnectEvent>(
+            'onWsReconnect',
+            {
+              group: `${group}.host`,
+              connectionAttempt: this._reconnectAttempts,
+              timeoutMs,
+            }
+          );
+
           setTimeout(() => {
             this.createWebSocket(); // start fresh with a new token
-          }, timeout);
+          }, timeoutMs);
         }
       });
 
       ws.addEventListener('error', (error) => {
+        if (this._heartbeatIntervalId) {
+          clearInterval(this._heartbeatIntervalId);
+        }
+
+        if (this._resyncIntervalId) {
+          clearInterval(this._resyncIntervalId);
+        }
+
         this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketErrorEvent>(
           'onWsError',
           {
             group: `${group}.host`,
             error: error as unknown as ErrorEvent,
+            connectionAttempt,
           }
         );
 
@@ -269,22 +370,15 @@ export class WeaveStoreAzureWebPubSubSyncHost {
           ws.close(); // ensure cleanup
         }
       });
-
-      setTimeout(() => {
-        if (ws.readyState === WebSocket.OPEN) {
-          ws.close();
-          this.server.emitEvent<WeaveStoreAzureWebPubsubOnWebsocketOnTokenRefreshEvent>(
-            'onWsTokenRefresh',
-            { group: `${group}.host` }
-          );
-          this.createWebSocket(); // start fresh with a new token
-        }
-      }, expirationTimeInMinutes * 0.75 * 60 * 1000);
     });
   }
 
   async start(): Promise<void> {
     await this.createWebSocket();
+  }
+
+  public isConnected(): boolean {
+    return this._conn && this._conn.readyState === WebSocket.OPEN;
   }
 
   async stop(): Promise<void> {